Refactored Design Smell in ArgumentMatchers.
